ORA Peer Assessment score override
This feature allows a course staff/instructor to override
the peer score, produced by the normal workflow, with a
manually input score.

The peer score can be overriden in 2 scanrios:
- The student has an existing peer score.
OR
- The due date for peer assessments has passed.
This prevents the override score from itself being overriden
by a peer score. Although, the peer score can be override
any number of times, only the last one will count.

This PR depends on the adjustments made to the Submissions
repo in PR 1
Stanford-Online/edx-submissions#1

The override score is stored in Submissions
as a score with a null submission and the original
peer score is treated as immutable. This was at the suggestion
of Dave O.

This PR provides context for the Submissions PR #1 above but the
process for merging these PR's needs to be:

- Review/merge Submissions PR
- Adjust commit hash appropriately of Submissions repo in base.txt
of this PR
- Review/merge this PR
- Adjust commit hashes for Submissions and ORA2 repos in a, yet
to be created, edx-platform PR.